# src/api/v1/public/swagger.yaml
openapi: 3.0.3
info:
  title: API Documentation Nhóm 2
  description: Chi tiết về API để implment cho dự án
  version: 1.0.0
servers:
  -  url: https://drpteam.vercel.app
  -  url: https://api-kyotrankmas-projects.vercel.app
...
paths:
...
  /v1/api/authenticate:
    get:
      tags:
        - users
      summary: Authen token cho user
      description: 'Xác thực access token cho mỗi phiên đăng nhập mà không cần đăng nhập bằng tài khoản, mật khẩu. Lưu ý: Cần truyền vào header Authorization với giá trị là Bearer + accessToken'
      operationId: authenToken
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Xác thực thành công"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/authenToken'
                  
        '400':
          description: "Accesstoken không chính xác hoặc hết hạn"
          content:
            application/json:
              schema: 
                type: object
                properties:
                  message:
                    type: string
                    example: Token not valid
        '401':
          description: "Accesstoken rỗng"
          content:
            application/json:
              schema: 
                type: object
                properties:
                  message:
                    type: string
                    example: Unauthorize user
...
components:
  schemas:
...
    authenToken: 
        type: object
        properties:
          message:
            type: string
            example: Authen successfully
          account:
            type: object
            properties:
              id: 
                type: string
                example: 663b48fc09800f0e9bb8e758
              username: 
                type: string
                example: Kyo
              email:
                type: string
                example: quangdieu10092@gmail.com
          apiKeyAIService:
            type: string
            example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
...
  securitySchemes:
    bearerAuth:          
      type: http
      scheme: bearer
      bearerFormat: JWT    